KB source formatting
git-svn-id: svn://svn.liferay.com/repos/public/plugins/trunk@75542 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portlets/knowledge-base-portlet/docroot/WEB-INF/src/com/liferay/knowledgebase/admin/portlet/AdminPortlet.java
@@ -200,7 +200,7 @@ public class AdminPortlet extends MVCPortlet {
 		throws PortletException, IOException {
 
 		try {
-			Integer status = WorkflowConstants.STATUS_ANY;
+			int status = WorkflowConstants.STATUS_ANY;
 
 			renderRequest.setAttribute(WebKeys.KNOWLEDGE_BASE_STATUS, status);
 
@@ -211,7 +211,7 @@ public class AdminPortlet extends MVCPortlet {
 
 			if (resourcePrimKey > 0) {
 				article = ArticleServiceUtil.getLatestArticle(
-					resourcePrimKey, status.intValue());
+					resourcePrimKey, status);
 			}
 
 			renderRequest.setAttribute(WebKeys.KNOWLEDGE_BASE_ARTICLE, article);
